# precompute tag to_string variants