--- a/lib/sneeze/internals.ex
+++ b/lib/sneeze/internals.ex
@@ -1,4 +1,33 @@
 defmodule Sneeze.Internal do
+  require Sneeze.Macros
+
+  Sneeze.Macros.define_tags_to_strings([
+    :a,
+    :span,
+    :li,
+    :div,
+    :td,
+    :br,
+    :ul,
+    :tr,
+    :link,
+    :p,
+    :title,
+    :script,
+    :style,
+    :button,
+    :label,
+    :code,
+    :h1,
+    :h2,
+    :h3,
+    :h4,
+    :h5,
+    :meta,
+    :body,
+    :head
+  ])
+
   def void_tags() do
     [
       :area,
@@ -29,25 +58,25 @@ defmodule Sneeze.Internal do
   end
 
   def render_opening_tag(tag_name) do
-    ["<", to_string(tag_name), ">"]
+    ["<", tag_to_string(tag_name), ">"]
   end
 
   def render_opening_tag(tag_name, attribs) do
     attrib_iolist = attributes_to_iolist(attribs)
-    ["<", to_string(tag_name), attrib_iolist, ">"]
+    ["<", tag_to_string(tag_name), attrib_iolist, ">"]
   end
 
   def render_closing_tag(tag_name) do
-    ["</", to_string(tag_name), ">"]
+    ["</", tag_to_string(tag_name), ">"]
   end
 
   def render_void_tag(tag_name) do
-    ["<", to_string(tag_name), " />"]
+    ["<", tag_to_string(tag_name), " />"]
   end
 
   def render_void_tag(tag_name, attribs) do
     attrib_iolist = attributes_to_iolist(attribs)
-    ["<", to_string(tag_name), attrib_iolist, " ", "/>"]
+    ["<", tag_to_string(tag_name), attrib_iolist, " ", "/>"]
   end
 
   def render_tag(tag) do
@@ -57,4 +86,5 @@ defmodule Sneeze.Internal do
   def render_tag(tag, attributes) do
     [render_opening_tag(tag, attributes), render_closing_tag(tag)]
   end
+
 end