--- a/lib/sneeze/internals.ex
+++ b/lib/sneeze/internals.ex
@@ -86,5 +86,4 @@ defmodule Sneeze.Internal do
   def render_tag(tag, attributes) do
     [render_opening_tag(tag, attributes), render_closing_tag(tag)]
   end
-
 end