--- a/lib/sneeze/macros.ex
+++ b/lib/sneeze/macros.ex
@@ -4,11 +4,11 @@ defmodule Sneeze.Macros do
   defmacro define_tags_to_strings(tags) do
     quote bind_quoted: [tags: tags] do
       Enum.each(tags, fn tagname ->
-         as_str = to_string(tagname)
+        as_str = to_string(tagname)
 
-         defp tag_to_string(unquote(tagname)) do
-           unquote(as_str)
-         end
+        defp tag_to_string(unquote(tagname)) do
+          unquote(as_str)
+        end
       end)
 
       defp tag_to_string(unknown_tag), do: to_string(unknown_tag)