@JuneKelly 2023-06-28
Thanks for this! I like where this is going.

On the question of how to build the list of tags, I guess wikipedia pages are as good as any as a data source. I get the impression that this macro won't add very much overhead to the runtime (all things considered) so we could err on the side of including more tags in the list.

One other thing: it looks like the CI build is failing on the `format` stage. Would you mind running `make format` and pushing the changes? Thanks!