@ckampfe 2023-06-28
I ran a quick script to compute a bit more data, and this is what it came up with:

```
at [ 17:32:41 ] ➜ ./aggregate_tags.exs
fetching https://elixir-lang.org/
fetching https://www.phoenixframework.org/
fetching https://www.nytimes.com/
fetched https://elixir-lang.org/
fetched https://www.phoenixframework.org/
fetched https://www.nytimes.com/
[
  {"div", 1278},
  {"span", 332},
  {"a", 252},
  {"li", 182},
  {"section", 123},
  {"p", 116},
  {"h3", 100},
  {"figure", 94},
  {"img", 82},
  {"svg", 75},
  {"picture", 51},
  {"script", 49},
  {"meta", 44},
  {"link", 43},
  {"button", 41},
  {"figcaption", 36},
  {"source", 29},
  {"noscript", 29},
  {"path", 21},
  {"polygon", 20},
  {"h4", 19},
  {"ul", 17},
  {"i", 14},
  {"code", 14},
  {"nav", 12}
]
```

I tweaked the tagset that gets generated fns in [5427145](https://github.com/JuneKelly/sneeze/pull/19/commits/54271450f7d1b028925dfc0e12ca3b8057a899d5) a bit in response to this data as well as [this](https://css-tricks.com/average-web-page-data-analyzing-8-million-websites/) survey of what tags pages tend to contain.

The script source used to generate the data:

```elixir
#!/usr/bin/env elixir

Mix.install([:req, :easyhtml, {:util, git: "https://github.com/ckampfe/util.git"}])

links = [
  "https://elixir-lang.org/",
  "https://www.phoenixframework.org/",
  "https://www.nytimes.com/",
]

links
|> Task.async_stream(fn link ->
  IO.puts("fetching #{link}")
  {link, Req.get!(link)}
end)
|> Enum.map(fn {:ok, {link, res}} ->
  IO.puts("fetched #{link}")
  Map.fetch!(res, :body)
end)
|> Enum.map(fn body ->
  EasyHTML.parse!(body)
end)
|> Enum.flat_map(fn parsed ->
  Util.traverse(List.first(parsed.nodes),
    fn
      {_tag, _attrs, children} when is_list(children) -> 
        true
      _ -> false
    end,
    fn {_tag, _attrs, children} -> children end
  )
end)
|> Enum.group_by(fn {tag, _, _} -> tag; _ -> nil end)
|> Enum.map(fn {tag, values} -> {tag, Enum.count(values)} end)
|> Enum.filter(fn {tag, _count} -> !is_nil(tag) end)
|> Enum.sort_by(fn {_tag, count} -> count end)
|> Enum.reverse()
|> Enum.take(25)
|> IO.inspect()
```