@JuneKelly 2023-07-10
This all looks great, thanks! 

I'm going to merge it today and release it in 2.0.0